Move license related APIs into an api module
