+++ lib/plottr_components/src/components/dashboard/account/Account.js
@@ -25,7 +25,6 @@ const AccountConnector = (connector) => {
   const ExpiredView = UnconnectedExpiredView(connector)
   const LicenseInfo = UnconnectedLicenseInfo(connector)
   const ProInfo = UnconnectedProInfo(connector)
-  const BetaInfo = UnconnectedBetaInfo(connector)
 
   // TODO: Move ALL of this logic into pltr/v2/selectors/license.js
   //
@@ -84,7 +83,6 @@ const AccountConnector = (connector) => {
       if (isInTrialModeWithExpiredTrial) return <ExpiredView darkMode={darkMode} />
 
       const body = []
-      if (os == 'unknown') body.push(<BetaInfo key="beta" />)
       if (hasCurrentProLicense) body.push(<ProInfo key="pro" />)
 
       if (hasLicense) {
