+++ lib/plottr_components/src/components/dashboard/navigation/DashboardBody.js
@@ -14,7 +14,7 @@ import { checkDependencies } from '../../checkDependencies'
 const DashboardBodyConnector = (connector) => {
   const {
     platform: {
-      license: { checkForActiveLicense, useTrialStatus },
+      license: { checkForActiveLicense },
       settings: { saveAppSetting },
       reloadMenu,
       os,
@@ -22,7 +22,6 @@ const DashboardBodyConnector = (connector) => {
   } = connector
   checkDependencies({
     checkForActiveLicense,
-    useTrialStatus,
     saveAppSetting,
     reloadMenu,
   })
