+++ lib/plottr_components/src/connections/pltr.js
@@ -145,7 +145,6 @@ const pltrTypeSpecs = {
   updateBeatHierarchyFlag: PropTypes.func.isRequired,
   license: PropTypes.shape({
     checkForActiveLicense: PropTypes.func.isRequired,
-    licenseStore: PropTypes.object.isRequired,
     verifyLicense: PropTypes.func.isRequired,
     trial90days: PropTypes.array.isRequired,
     checkForPro: PropTypes.func.isRequired,
@@ -210,7 +209,6 @@ const pltrTypeSpecs = {
   lockRCE: PropTypes.func.isRequired,
   machineIdSync: PropTypes.func.isRequired,
   extractImages: PropTypes.func.isRequired,
-  useProLicenseInfo: PropTypes.func.isRequired,
   firebase: PropTypes.shape({
     startUI: PropTypes.func.isRequired,
     firebaseUI: PropTypes.func.isRequired,
