+++ src/api/index.js
@@ -1,4 +1,5 @@
 import * as fileSystemAPIs from './file-system-apis'
 import * as firebaseAPIs from './firebase-apis'
+import * as licenseServerAPIs from './license-server-apis'
 
-export { fileSystemAPIs, firebaseAPIs }
+export { fileSystemAPIs, firebaseAPIs, licenseServerAPIs }
