--- src/common/licensing/special_codes.js
@@ -1,2 +1,0 @@
-export const trial90days = ['nanoCAMP@90', 'infoSTACK90!']
-export const trial60days = ['infoSTACK60!']
